+++ Tests/SwiftLintFrameworkTests/ReporterTests.swift
@@ -17,7 +17,8 @@ class ReporterTests: XCTestCase {
             SonarQubeReporter.self,
             MarkdownReporter.self,
             GitHubActionsLoggingReporter.self,
-            GitLabJUnitReporter.self
+            GitLabJUnitReporter.self,
+            RelativePathReporter.self
         ]
         for reporter in reporters {
             XCTAssertEqual(reporter.identifier, reporterFrom(identifier: reporter.identifier).identifier)
@@ -135,5 +136,23 @@ class ReporterTests: XCTestCase {
         let expectedOutput = stringFromFile("CannedMarkdownReporterOutput.md")
         let result = MarkdownReporter.generateReport(generateViolations())
         XCTAssertEqual(result, expectedOutput)
+    }
+
+    func testRelativePathReporter() {
+        let expectedOutput = stringFromFile("CannedRelativePathReporterOutput.txt")
+        let result = RelativePathReporter.generateReport(generateViolations())
+        XCTAssertEqual(result, expectedOutput)
+    }
+
+    func testRelativePathReporterPaths() {
+        let relativePath = "filename"
+        let absolutePath = FileManager.default.currentDirectoryPath + "/" + relativePath
+        let location = Location(file: absolutePath, line: 1, character: 2)
+        let violation = StyleViolation(ruleDescription: LineLengthRule.description,
+                                       location: location,
+                                       reason: "Violation Reason")
+        let result = RelativePathReporter.generateReport([violation])
+        XCTAssertFalse(result.contains(absolutePath))
+        XCTAssertTrue(result.contains(relativePath))
     }
 }