+++ CHANGELOG.md
@@ -50,6 +50,10 @@
   `quick_discouraged_call`.  
   [David Steinacher](https://github.com/stonko1994)
   [#4626](https://github.com/realm/SwiftLint/issues/4626)
+  
+* Add `relative-path` reporter to generate reports with relative file paths.  
+  [Roya1v](https://github.com/roya1v)
+  [#4660](https://github.com/realm/SwiftLint/issues/4660)
 
 * Let `number_separator` rule trigger on misplaced separators, e.g. `10_00`.  
   [SimplyDanny](https://github.com/SimplyDanny)